refactor: вынести вычисление размера экрана в хук useCountMovies
